requirements.txt
requests==2.31.0
parsel==1.8.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
reportlab==4.0.7
spacy>=3.7.0